refactor: MemberSessionService 인터페이스로 변경
